chore(data-table): restructuring story directory
