+++ packages/carbon-web-components/src/components/data-table/stories/expansion/data-table-expansion-story.ts
@@ -9,14 +9,14 @@
 
 import { html } from 'lit';
 import { boolean, select, text } from '@storybook/addon-knobs';
-import { prefix } from '../../../globals/settings';
-import { TABLE_SIZE } from '../table';
+import { prefix } from '../../../../globals/settings';
+import { TABLE_SIZE } from '../../table';
 // @ts-ignore
 import Download16 from '@carbon/web-components/es/icons/download/16';
 // @ts-ignore
 import Settings16 from '@carbon/web-components/es/icons/settings/16';
-import '../index';
-import storyDocs from '../data-table-story.mdx';
+import '../../index';
+import storyDocs from './data-table-story.mdx';
 
 const sizes = {
   [`xs (${TABLE_SIZE.XS})`]: TABLE_SIZE.XS,
