+++ packages/carbon-web-components/src/components/data-table/stories/sorting/data-table-sorting-story.ts
@@ -9,17 +9,12 @@
 
 import { html } from 'lit';
 import { boolean, select, text } from '@storybook/addon-knobs';
-import { prefix } from '../../../globals/settings';
-import { TABLE_SIZE } from '../table';
-import Add from '@carbon/web-components/es/icons/add/16';
-import Save from '@carbon/web-components/es/icons/save/16';
-import TrashCan from '@carbon/web-components/es/icons/trash-can/16';
-// @ts-ignore
-import Download16 from '@carbon/web-components/es/icons/download/16';
+import { prefix } from '../../../../globals/settings';
+import { TABLE_SIZE } from '../../table';
 // @ts-ignore
 import Settings16 from '@carbon/web-components/es/icons/settings/16';
-import '../index';
-import storyDocs from '../data-table-story.mdx';
+import '../../index';
+import storyDocs from './data-table-story.mdx';
 
 const sizes = {
   [`xs (${TABLE_SIZE.XS})`]: TABLE_SIZE.XS,
@@ -29,109 +24,16 @@ const sizes = {
   [`xl (${TABLE_SIZE.XL})`]: TABLE_SIZE.XL,
 };
 
-let headerCount = 6;
-let rowCount = 1;
-
-const insertInRandomPosition = (array, element) => {
-  const index = Math.floor(Math.random() * (array.length + 1));
-  return [...array.slice(0, index), element, ...array.slice(index)];
-};
-
-const addRow = () => {
-  const newRow = document.createElement('cds-table-row');
-
-  const templateRow = {
-    name: `New Row ${rowCount}`,
-    protocol: 'HTTP',
-    port: rowCount * 100,
-    rule: rowCount % 2 === 0 ? 'Round robin' : 'DNS delegation',
-    attached_groups: `Row ${rowCount}'s VM Groups`,
-    status: 'Starting',
-  };
-
-  for (let key in templateRow) {
-    if (Object.prototype.hasOwnProperty.call(templateRow, key)) {
-      const cell = document.createElement('cds-table-cell');
-      cell.textContent = templateRow[key];
-      newRow.appendChild(cell);
-    }
-  }
-
-  const rows = document.querySelectorAll('cds-table-row');
-  const diff = headerCount - Object.keys(templateRow).length;
-
-  [...Array(diff)].forEach(() => {
-    const newCell = document.createElement('cds-table-cell');
-    newCell.textContent = `Header ${headerCount - 1}`;
-    newRow.appendChild(newCell);
-  });
-
-  newRow.setAttribute('selection-name', `${rows.length}`);
-
-  const updatedRows = insertInRandomPosition([...rows], newRow);
-  updatedRows.forEach((e) => {
-    document.querySelector('cds-table-body')!.insertBefore(e, null);
-  });
-
-  rowCount++;
-};
-
-const addHeader = () => {
-  const headerRow = document.querySelector('cds-table-header-row');
-  const newHeader = document.createElement('cds-table-header-cell');
-  newHeader.textContent = `Header ${headerCount}`;
-  headerRow?.appendChild(newHeader);
-
-  const rows = document.querySelectorAll('cds-table-row');
-  rows.forEach((e) => {
-    const newCell = document.createElement('cds-table-cell');
-    newCell.textContent = `Header ${headerCount}`;
-    e.appendChild(newCell);
-  });
-  headerCount++;
-};
-
 export const Default = () => {
   return html`
-    <cds-table>
+    <cds-table is-sortable>
       <cds-table-header-title slot="title">DataTable</cds-table-header-title>
       <cds-table-header-description slot="description"
-        >Use the toolbar menu to add rows and
-        headers</cds-table-header-description
+        >With filtering</cds-table-header-description
       >
 
-      <cds-table-toolbar slot="toolbar">
-        <cds-table-batch-actions ?active="true">
-          <cds-button tooltip-position="bottom" tooltip-text="Add"
-            >${Add({ slot: 'icon' })}</cds-button
-          >
-          <cds-button tooltip-position="bottom" tooltip-text="Save"
-            >${Save({ slot: 'icon' })}</cds-button
-          >
-          <cds-button>${Save({ slot: 'icon' })}</cds-button>
-          <cds-button href="javascript:void 0" download="table-data.json">
-            Download ${Download16({ slot: 'icon' })}
-          </cds-button>
-        </cds-table-batch-actions>
-        <cds-table-toolbar-content ?has-batch-actions="true">
-          <cds-table-toolbar-search
-            placeholder="Filter table"></cds-table-toolbar-search>
-          <cds-overflow-menu toolbar-action>
-            ${Settings16({ slot: 'icon' })}
-            <cds-overflow-menu-body>
-              <cds-overflow-menu-item @click=${addRow}
-                >Add row</cds-overflow-menu-item
-              >
-              <cds-overflow-menu-item @click=${addHeader}
-                >Add header</cds-overflow-menu-item
-              >
-            </cds-overflow-menu-body>
-          </cds-overflow-menu>
-        </cds-table-toolbar-content>
-      </cds-table-toolbar>
-
       <cds-table-head>
-        <cds-table-header-row selection-name="header">
+        <cds-table-header-row>
           <cds-table-header-cell>Name</cds-table-header-cell>
           <cds-table-header-cell>Protocol</cds-table-header-cell>
           <cds-table-header-cell>Port</cds-table-header-cell>
@@ -141,7 +43,7 @@ export const Default = () => {
         </cds-table-header-row>
       </cds-table-head>
       <cds-table-body>
-        <cds-table-row selection-name="0">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 3</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>3000</cds-table-cell>
@@ -151,7 +53,7 @@ export const Default = () => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-row selection-name="1">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 1</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>443</cds-table-cell>
@@ -159,7 +61,7 @@ export const Default = () => {
           <cds-table-cell>Maureen's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-row selection-name="2">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 2</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>80</cds-table-cell>
@@ -167,7 +69,7 @@ export const Default = () => {
           <cds-table-cell>Andrew's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-row selection-name="3">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 6</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>3000</cds-table-cell>
@@ -177,7 +79,7 @@ export const Default = () => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-row selection-name="4">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 4</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>443</cds-table-cell>
@@ -185,7 +87,7 @@ export const Default = () => {
           <cds-table-cell>Mel's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-row selection-name="5">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 5</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>80</cds-table-cell>
@@ -202,7 +104,6 @@ export const Playground = (args) => {
   const {
     isSortable,
     locale,
-    radio,
     size,
     stickyHeader,
     useStaticWidth,
@@ -210,47 +111,19 @@ export const Playground = (args) => {
   } = args?.[`${prefix}-table`] ?? {};
   return html`
     <cds-table
-      expandable
       ?is-sortable=${isSortable}
       locale="${locale}"
-      ?radio=${radio}
       size="${size}"
       ?sticky-header="${stickyHeader}"
       ?use-static-width="${useStaticWidth}"
       ?use-zebra-styles="${useZebraStyles}">
       <cds-table-header-title slot="title">DataTable</cds-table-header-title>
       <cds-table-header-description slot="description"
-        >Use the toolbar menu to add rows and
-        headers</cds-table-header-description
+        >With filtering</cds-table-header-description
       >
 
-      <cds-table-toolbar slot="toolbar">
-        <cds-table-batch-actions ?active="true">
-          <cds-button>Delete ${TrashCan({ slot: 'icon' })}</cds-button>
-          <cds-button>Save ${Save({ slot: 'icon' })}</cds-button>
-          <cds-button href="javascript:void 0" download="table-data.json">
-            Download ${Download16({ slot: 'icon' })}
-          </cds-button>
-        </cds-table-batch-actions>
-        <cds-table-toolbar-content ?has-batch-actions="true">
-          <cds-table-toolbar-search
-            placeholder="Filter table"></cds-table-toolbar-search>
-          <cds-overflow-menu toolbar-action>
-            ${Settings16({ slot: 'icon' })}
-            <cds-overflow-menu-body>
-              <cds-overflow-menu-item @click=${addRow}
-                >Add row</cds-overflow-menu-item
-              >
-              <cds-overflow-menu-item @click=${addHeader}
-                >Add header</cds-overflow-menu-item
-              >
-            </cds-overflow-menu-body>
-          </cds-overflow-menu>
-        </cds-table-toolbar-content>
-      </cds-table-toolbar>
-
       <cds-table-head>
-        <cds-table-header-row selection-name="header">
+        <cds-table-header-row>
           <cds-table-header-cell>Name</cds-table-header-cell>
           <cds-table-header-cell>Protocol</cds-table-header-cell>
           <cds-table-header-cell>Port</cds-table-header-cell>
@@ -260,7 +133,7 @@ export const Playground = (args) => {
         </cds-table-header-row>
       </cds-table-head>
       <cds-table-body>
-        <cds-table-row selection-name="0">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 3</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>3000</cds-table-cell>
@@ -270,11 +143,7 @@ export const Playground = (args) => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content 3</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row selection-name="1">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 1</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>443</cds-table-cell>
@@ -282,11 +151,7 @@ export const Playground = (args) => {
           <cds-table-cell>Maureen's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row selection-name="2">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 2</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>80</cds-table-cell>
@@ -294,11 +159,7 @@ export const Playground = (args) => {
           <cds-table-cell>Andrew's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row selection-name="3">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 6</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>3000</cds-table-cell>
@@ -308,11 +169,7 @@ export const Playground = (args) => {
             ><cds-link disabled>Disabled</cds-link></cds-table-cell
           >
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row selection-name="4">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 4</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>443</cds-table-cell>
@@ -320,11 +177,7 @@ export const Playground = (args) => {
           <cds-table-cell>Mel's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Starting</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
-        <cds-table-row selection-name="5">
+        <cds-table-row>
           <cds-table-cell>Load Balancer 5</cds-table-cell>
           <cds-table-cell>HTTP</cds-table-cell>
           <cds-table-cell>80</cds-table-cell>
@@ -332,10 +185,6 @@ export const Playground = (args) => {
           <cds-table-cell>Ronja's VM Groups</cds-table-cell>
           <cds-table-cell><cds-link>Active</cds-link></cds-table-cell>
         </cds-table-row>
-        <cds-table-expanded-row>
-          <h6>Expandable row content</h6>
-          <div>Description here</div>
-        </cds-table-expanded-row>
       </cds-table-body>
     </cds-table>
   `;
@@ -344,9 +193,8 @@ export const Playground = (args) => {
 Playground.parameters = {
   knobs: {
     [`${prefix}-table`]: () => ({
-      isSortable: boolean('Is sortable', false),
+      isSortable: boolean('Is sortable', true),
       locale: text('Locale', 'en'),
-      radio: boolean('Radio', false),
       size: select('Size', sizes, TABLE_SIZE.LG),
       stickyHeader: boolean('Sticky header', false),
       useStaticWidth: boolean('Use static width', false),
@@ -356,7 +204,7 @@ Playground.parameters = {
 };
 
 export default {
-  title: 'Components/DataTable/Dynamic',
+  title: 'Components/DataTable/Sorting',
   parameters: {
     ...storyDocs.parameters,
   },
